+++ src/components/utils/camera.js
@@ -1,13 +1,13 @@
 import { canvasToBlob } from './blob'
 
-export const screenshot = (webcam, callback) => {
+export const screenshot = (webcam, callback, mimeType) => {
   const canvas = webcam && webcam.getCanvas()
-  if (!canvas){
+  if (!canvas) {
     console.error('webcam canvas is null')
     return
   }
   const sdkMetadata = getDeviceInfo(webcam.stream)
-  canvasToBlob(canvas, blob => callback(blob, sdkMetadata))
+  canvasToBlob(canvas, blob => callback(blob, sdkMetadata), mimeType)
 }
 
 export const getRecordedVideo = (webcam, callback) => {
@@ -16,11 +16,11 @@ export const getRecordedVideo = (webcam, callback) => {
   callback({ blob, sdkMetadata })
 }
 
-const getDeviceInfo = (stream) => {
+const getDeviceInfo = stream => {
   if (stream) {
     const videoTrack = stream.getVideoTracks()[0] || {}
     const audioTrack = stream.getAudioTracks()[0] || {}
-    return {camera_name: videoTrack.label, microphone_name: audioTrack.label}
+    return { camera_name: videoTrack.label, microphone_name: audioTrack.label }
   }
   return {}
 }