+++ test/pageobjects/CrossDeviceMobileNotificationSent.js
@@ -1,20 +1,18 @@
 import BasePage from './BasePage.js'
 import { verifyElementCopy } from '../utils/mochaw'
 import { asyncForEach } from '../utils/async'
-import { By } from 'selenium-webdriver'
 
 class CrossDeviceMobileNotificationSent extends BasePage {
   get submessage() { return this.$('.onfido-sdk-ui-crossDevice-MobileNotificationSent-submessage')}
   get mayTakeFewMinutesMessage() { return this.$('.onfido-sdk-ui-crossDevice-MobileNotificationSent-boldMessage')}
-  get yourMobilePhoneIconSelector() { return By.css('.onfido-sdk-ui-Theme-icon')}
-  get yourMobilePhoneIcon() { return this.$('.onfido-sdk-ui-Theme-icon')}
+  async yourMobilePhoneIcon() { return this.waitAndFind('.onfido-sdk-ui-Theme-icon')}
   get tipsHeader() { return this.$('.onfido-sdk-ui-Theme-header')}
   get tips() { return this.$('.onfido-sdk-ui-Theme-helpList li')}
   get resendLink() { return this.$('.onfido-sdk-ui-crossDevice-MobileNotificationSent-cancel')}
 
   async verifyTitle(copy) {
     const mobileNotificationSentStrings = copy.cross_device
-    verifyElementCopy(this.title, mobileNotificationSentStrings.mobile_notification_sent.title)
+    verifyElementCopy(this.title(), mobileNotificationSentStrings.mobile_notification_sent.title)
   }
 
   async verifySubmessage(copy) {
@@ -27,7 +25,7 @@ class CrossDeviceMobileNotificationSent extends BasePage {
   }
 
   async verifycrossDeviceCheckYourMobilePhoneIcon() {
-    this.yourMobilePhoneIcon.isDisplayed()
+    this.yourMobilePhoneIcon().isDisplayed()
   }
 
   async verifyTipsHeader(copy) {