+++ test/pageobjects/DocumentUpload.js
@@ -1,25 +1,23 @@
 import BasePage from './BasePage.js'
 const path = require('path')
 import { verifyElementCopy } from '../utils/mochaw'
-import { By } from 'selenium-webdriver'
 
 class DocumentUpload extends BasePage {
-  get crossDeviceIcon() { return this.$('.onfido-sdk-ui-crossDevice-SwitchDevice-icon')}
+  async crossDeviceIcon() { return this.waitAndFind('.onfido-sdk-ui-crossDevice-SwitchDevice-icon')}
   get crossDeviceHeader() { return this.$('.onfido-sdk-ui-crossDevice-SwitchDevice-header')}
   get crossDeviceSubMessage() { return this.$('.onfido-sdk-ui-crossDevice-SwitchDevice-submessage')}
   get crossDeviceArrow() { return this.$('.onfido-sdk-ui-crossDevice-SwitchDevice-chevron')}
   get uploaderIcon() { return this.$('.onfido-sdk-ui-Theme-icon')}
-  get uploaderInstructionsMessageSelector() { return By.css('.onfido-sdk-ui-Uploader-instructionsCopy')}
-  get uploaderInstructionsMessage() { return this.$('.onfido-sdk-ui-Uploader-instructionsCopy')}
+  async uploaderInstructionsMessage() { return this.waitAndFind('.onfido-sdk-ui-Uploader-instructionsCopy')}
   get uploaderBtn() { return this.$('.onfido-sdk-ui-Uploader-buttons')}
-  getUploadInput() { return (async ()=>{
-    const input = this.$('.onfido-sdk-ui-CustomFileInput-input')
-    // eslint-disable-next-line prefer-arrow-callback
-    this.driver.executeScript(function(el) {
+  async uploadInput() { return this.waitAndFind('.onfido-sdk-ui-CustomFileInput-input') }
+  async getUploadInput() {
+    const input = this.uploadInput()
+    this.driver.executeScript((el) => {
       el.setAttribute('style','display: block !important')
-    },input)
+    }, input)
     return input
-  })()}
+  }
 
   upload(filename) {
     const input = this.$('.onfido-sdk-ui-CustomFileInput-input')
@@ -30,7 +28,7 @@ class DocumentUpload extends BasePage {
 
   async verifyCrossDeviceUIElements(copy) {
     const documentUploadCrossDeviceStrings = copy.cross_device.switch_device
-    this.crossDeviceIcon.isDisplayed()
+    this.crossDeviceIcon().isDisplayed()
     verifyElementCopy(this.crossDeviceHeader, documentUploadCrossDeviceStrings.header)
     verifyElementCopy(this.crossDeviceSubMessage, documentUploadCrossDeviceStrings.submessage)
     this.crossDeviceArrow.isDisplayed()
@@ -47,62 +45,62 @@ class DocumentUpload extends BasePage {
 
   async verifyPassportTitle(copy) {
     const documentUploadStrings = copy.capture
-    verifyElementCopy(this.title, documentUploadStrings.passport.front.title)
+    verifyElementCopy(this.title(), documentUploadStrings.passport.front.title)
   }
 
   async verifyPassportInstructionMessage(copy) {
     const documentUploadStrings = copy.capture
-    verifyElementCopy(this.uploaderInstructionsMessage, documentUploadStrings.passport.front.instructions)
+    verifyElementCopy(this.uploaderInstructionsMessage(), documentUploadStrings.passport.front.instructions)
   }
 
   async verifyFrontOfDrivingLicenceTitle(copy) {
     const documentUploadStrings = copy.capture
-    verifyElementCopy(this.title, documentUploadStrings.driving_licence.front.title)
+    verifyElementCopy(this.title(), documentUploadStrings.driving_licence.front.title)
   }
 
   async verifyFrontOfDrivingLicenceInstructionMessage(copy) {
     const documentUploadStrings = copy.capture
-    verifyElementCopy(this.uploaderInstructionsMessage, documentUploadStrings.driving_licence.front.instructions)
+    verifyElementCopy(this.uploaderInstructionsMessage(), documentUploadStrings.driving_licence.front.instructions)
   }
 
   async verifyBackOfDrivingLicenceTitle(copy) {
     const documentUploadStrings = copy.capture
-    verifyElementCopy(this.title, documentUploadStrings.driving_licence.back.title)
+    verifyElementCopy(this.title(), documentUploadStrings.driving_licence.back.title)
   }
 
   async verifyBackOfDrivingLicenceInstructionMessage(copy) {
     const documentUploadStrings = copy.capture
-    verifyElementCopy(this.uploaderInstructionsMessage, documentUploadStrings.driving_licence.back.instructions)
+    verifyElementCopy(this.uploaderInstructionsMessage(), documentUploadStrings.driving_licence.back.instructions)
   }
 
   async verifyFrontOfIdentityCardTitle(copy) {
     const documentUploadStrings = copy.capture
-    verifyElementCopy(this.title, documentUploadStrings.national_identity_card.front.title)
+    verifyElementCopy(this.title(), documentUploadStrings.national_identity_card.front.title)
   }
 
   async verifyFrontOfIdentityCardInstructionMessage(copy) {
     const documentUploadStrings = copy.capture
-    verifyElementCopy(this.uploaderInstructionsMessage, documentUploadStrings.national_identity_card.front.instructions)
+    verifyElementCopy(this.uploaderInstructionsMessage(), documentUploadStrings.national_identity_card.front.instructions)
   }
 
   async verifyBackOfIdentityCardTitle(copy) {
     const documentUploadStrings = copy.capture
-    verifyElementCopy(this.title, documentUploadStrings.national_identity_card.back.title)
+    verifyElementCopy(this.title(), documentUploadStrings.national_identity_card.back.title)
   }
 
   async verifyBackOfIdentityCardInstructionMessage(copy) {
     const documentUploadStrings = copy.capture
-    verifyElementCopy(this.uploaderInstructionsMessage, documentUploadStrings.national_identity_card.back.instructions)
+    verifyElementCopy(this.uploaderInstructionsMessage(), documentUploadStrings.national_identity_card.back.instructions)
   }
 
   async verifySelfieUploadTitle(copy) {
     const documentUploadStrings = copy.capture
-    verifyElementCopy(this.title, documentUploadStrings.face.upload_title)
+    verifyElementCopy(this.title(), documentUploadStrings.face.upload_title)
   }
 
   async verifySelfieUploadInstructions(copy) {
     const documentUploadStrings = copy.capture
-    verifyElementCopy(this.uploaderInstructionsMessage, documentUploadStrings.face.instructions)
+    verifyElementCopy(this.uploaderInstructionsMessage(), documentUploadStrings.face.instructions)
   }
 }